Alpha Kafka support
Adds a Kafka event monitor and replaces the Ceilometer event
monitor as the default for OpenStack.

The event monitor subscribes to all topics and processes messages
from the Kafka bus in batches using the Consumer interface
from the ruby-kafka gem.

Seed brokers are assumed to be on localhost. See list of TODOs in
the event monitor code for other gotchas.